Comment out the light updating on another thread. Caused an issue with dark spots.
Issue probably occurred due to chunks being null during the lighting queue update, meaning that the light change was removed from the queue but not processed at all because its chunk was null. A possible solution could be to keep a queue of updates that didn't work out and throw those in the queue after for processing the next time.